Refactor code for generating speech #1
