+++ .env.example
@@ -1,3 +1,6 @@
 OPENAI_API_KEY=
 LINE_CHANNEL_ACCESS_TOKEN=
 LINE_CHANNEL_SECRET=
+GOOGLE_CSE_ID=
+GOOGLE_API_KEY=
+HOSTNAME=
